Clearer URLs and stricter authentication
