+++ Task04/config/auth.json
@@ -9,15 +9,15 @@
 		"/deploy": ["POST"]
 	},
 	"guest": {
-		"/ln(.*)": ["GET"],
-		"/viz(.*)": ["GET"],
-		"/css(.*)": ["GET"],
-		"/js(.*)": ["GET"],
-		"/img(.*)": ["GET"],
-		"/": ["GET"],
-		"/documentation(.*)": ["GET"],
+		"/documentation/(.*)": ["GET"],
 		"/institutes(.*)": ["GET"],
 		"/gateways(.*)": ["GET"],
-		"/sensors(.*)": ["GET"]
+		"/sensors(.*)": ["GET"],
+		"/graphs/(.*)": ["GET"],
+		"/img/(.*)": ["GET"],
+		"/css/(.*)": ["GET"],
+		"/js/(.*)": ["GET"],
+		"/legal": ["GET"],
+		"/": ["GET"]
 	}
 }
