+++ Task04/views/graphs/single.hbs
