+++ Task04/views/layout.hbs
@@ -31,10 +31,10 @@
 						<li class="dropdown">
 							<a href="javascript:void(0)" class="dropbtn">Climate</a>
 							<div class="dropdown-content">
-								<a href="/viz/network">Chart: network topology</a>
-								<a href="/viz/sensor">Chart: single sensor</a>
-								<a href="/viz/sensormult">Chart: multiple sensors in one site</a>
-								<a href="/viz/sensorcomp">Chart: comparison between sites</a>
+								<a href="/graphs/network">Chart: network topology</a>
+								<a href="/graphs/single">Chart: single sensor</a>
+								<a href="/graphs/multiple">Chart: multiple sensors in one site</a>
+								<a href="/graphs/sites">Chart: comparison between sites</a>
 							</div>
 						</li>
 						<li class="dropdown">
@@ -71,7 +71,7 @@
 
 		<div class="footer">
 			<p>
-				This projet is funded by ANR grant number ANR-19-CE32-0001.<br> <a href="/ln">Legal notice</a>
+				This projet is funded by ANR grant number ANR-19-CE32-0001.<br> <a href="/legal">Legal notice</a>
 			</p>
 		</div>
 	</body>
