+++ Task04/views/legal.hbs
